Update patch set 1
Patch Set 1: Code-Review+2

LGTM with one reservation.

Long-term, we'll want to combine condition-based propagation with definition-based propagation, and I'm not sure we can do this with a simple tree walk (not sure means, really, not sure -- I know how to do it with a sparse-tree-lookup data structure, but maybe there are clever tree-walk tricks that work just as well).

But a bird in the hand is a darn good thing.  If we keep tree-walk-style analysis for nilcheckelim, we'll need to be ready to generalize it for array bounds checks and type checks.

I've got a sparse lookup data structure half-done (red-black trees is the largest part, the rest is fenceposts); one annoying thing I noticed is that it also fails to do well for tiny trees.  Do we plan to keep the simple algorithms and enable them whenever the number of blocks is below some small threshold?  I had also intended to instrument a compiler to collect a distribution of block sizes and perhaps block depths.

Patch-set: 1
Label: Code-Review=+2