Update patch set 2
Patch Set 2:

(3 comments)

I'm happy to just add a comment and switch back to the patch #1, but here's an attempt at correcting the removal of nil checks on the false branch.  The downside is it relies on the nilcheck node being a dominator for the false side though.

Patch-set: 2
Label: Code-Review=0