runtime: cache stack roots for sleeping goroutines between GCs
When scanning, get around having to traceback stacks by caching the
pointers encountered into the memory between the top of the stack
and the end of the stack allocation. This lowers cpu time spent for
the scan phase.

DO NOT SUBMIT

Clearly meant for go1.6. This is a reroll of
https://go-review.googlesource.com/#/c/3030/

All Gs have their pointers cached right now, which results in a net
loss for batch operation, particularly running all.bash. Need to
figure out heuristics for when we enable caching.

Need proper benchmarks.

Need to validate assertions made in comments about soundness. syncsend
trick scares me a bit

Change-Id: Idef30f85b0e327df8e15f4f7fe5abd4f23aef1a3