Update patch set 8
Patch Set 8: Code-Review+1

I must admit that at this point I am a little unsure that the original algorithm was correct.
I need to read this one more carefully, but I've been noodling around on my own and very nicely put my foot in a big steaming pile of incorrect assumptions.

As an example of "how was this right in the original?":

  if x != nil {
    // x is not nil
  } else {
    // x is nil
  }
  // x is ?? but this block is dominated by the test?

So I am going to stare at this and at my own code some more.

Patch-set: 8