Update patch set 8
Patch Set 8:

I have a change in mind, but I am not sure it is "better".
I'm a little worried about compiler speed at this point; look at the pushback we're getting on 1.5.

I need to re-review yours; I see no reason not to commit a performance improvement, and then we'll turn the crank again on trying to make it more general, and see if that actually is worth the cost.

Patch-set: 8