Update patch set 9
Patch Set 9:

Modulo Josh's most recent comments, I'm happy.  This seems like a good bird-in-the-hand, though I am not sure it will generalize for bounds-checking in the style from the paper Josh mentioned off in email ( http://homepages.dcc.ufmg.br/~fernando/publications/papers/SBLP2011_douglas.pdf )

That requires that you reify the "different" versions of a value in the two branches following an if, but I don't think we're up for doing it via adding yet more nodes to the SSA graph in 1.6 (I did something like this in years past, then you have to get rid of all the excess nodes after you are through with them).  I'm starting to think we could do a simple forward version in a tree-walk, with the same "critical-edges-removed would sure be nice" constraint that the sparse version would have.  But, best way to play with it is to have it in the code base.

I may just use the sparse framework to test "should we do more?" because I don't think it is as fast as this one, but it is somewhat more flexible.  I'll check, of course.

Patch-set: 9