Update patch set 5
Patch Set 5:

FWIW I recreated this as CL 12747, for issue #11886.
It would be nice to understand the runtime/pprof behavior if it is still broken, but flaky profiling is still an improvement on 'everything crashes'. If we can't figure out it we can roll it back.

Patch-set: 5
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0