Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1

This CL is missing the changes in ssa/gen.
You'd also need changes to ssa.go to actually implement the xor (something with OpAMD64XORQ & friends).

Could you hold off on this?  I have a CL almost ready which reorgs some of this code and will handle the issues above.

Patch-set: 2