Update patch set 2
Patch Set 2:

LGTM for the experiment.

I don't remember if I ever worked out what was going on here. But if I had to guess now, every time I see a list of */arm failures that excludes linux/arm and includes atomic operations, I assume it's http://golang.org/issue/7977

That is something I was hoping to fix for 1.5 but ran out of time before the freeze. The problem, as I understand it, is our CAS implementation is inadequate on ARMv7. It works on some common hardware, but not all. (I have a failing iPad on my desk, see #10189.)

Restricting ourselves to GOARM=7, I believe we can solve this the way minux suggested on the issue by adding a DMB instruction. But I thought it was too late back at the 1.5 freeze, so I assume it is far too late now.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0