Update patch set 1
Patch Set 1:

Again, for discussion. This was a TODO, so I implemented it to see what it looked like. It seems ok, although it is unfortunate that some of the rewrite go back and forth repeatedly from type to type index.

I didn't notice any obvious performance differences, but I didn't go searching for them either.

Patch-set: 1