Update patch set 1
Patch Set 1:

> Also, can we implement type-ish methods (IsPointer,...) on the indexes directly?  They would need to get the func ptr somehow to look up the info.

What do you think about making the types array a package level var?

It's a bit odd, and not concurrency-safe, but (a) that would make it easy to implement type-ish methods on the indices and (b) in a given package, we are going to encounter the same types repeatedly, so it is probably good for performance not to tear it down and recreate it for each function.

Patch-set: 1