Update patch set 1
Patch Set 1:

> > Also, can we implement type-ish methods (IsPointer,...) on the
 > indexes directly?  They would need to get the func ptr somehow to
 > look up the info.
 > 
 > What do you think about making the types array a package level var?
 > 
 > It's a bit odd, and not concurrency-safe, but (a) that would make
 > it easy to implement type-ish methods on the indices and (b) in a
 > given package, we are going to encounter the same types repeatedly,
 > so it is probably good for performance not to tear it down and
 > recreate it for each function.

That would mean at most 1<<32 types in a whole package, instead of just a function.  I guess I'm ok with that.

The package-level var would need synchronization, because potentially at least we'd want to run compiles in parallel.  We could put a per-function cache in front of it if needed (or just a cache in the ssa building state, where we'd do most of the type lookups).

Devils advocate, why wouldn't we do this for the whole compiler instead of just SSA?

Patch-set: 1