crypto/cipher: use amd64 acceleration for GHASH computation.
This doubles the throughput of AES-GCM encryption and decryption, mostly
due to being much faster to perform the GF(2^128) multiplication. The
implementation of the multiplication step is based on the gfmul function
from this Intel whitepaper:

    https://software.intel.com/sites/default/files/managed/72/cc/clmul-wp-rev-2.02-2014-04-20.pdf

I suspect we could do better by using the techniques in that paper for
the overall AES-GCM process, but this gives a nice win already.

For #11929.

Before:
    BenchmarkAESGCMSeal1K-12          200000              8449 ns/op         121.19 MB/s
    BenchmarkAESGCMOpen1K-12          200000              8461 ns/op         121.01 MB/s

After:
    BenchmarkAESGCMSeal1K-12          300000              4028 ns/op         254.18 MB/s
    BenchmarkAESGCMOpen1K-12          300000              4047 ns/op         252.97 MB/s

Change-Id: I819339be142e67d3482b832ff012afe036e96222