Update patch set 2
Patch Set 2: Run-TryBot+1

PTAL

The changes in this last round are due to the fact that in the old code 'go build -o foo.dylib -buildmode c-shared ...' actually invoked the linker with -o foo.dylib (no intermediate temporary file), and now that -o uses a proper install there is a temporary file that is copied to foo.dylib instead. But on OS X the linker records the -o argument in the actual binary, so copying from /tmp/blah/a.out to foo.dylib and linking against foo.dylib still produces a binary looking for /tmp/blah/a.out! I threaded the name down so that the linker writes to /tmp/blah/foo.dylib while in /tmp/blah as plain foo.dylib. This makes the file more "portable". This was all accidentally working in the test, but would have failed for any real 'go install' writing to an installed location.

Patch-set: 2
Label: Run-TryBot=+1