Update patch set 1
Patch Set 1:

I introduced this bug in https://go-review.googlesource.com/12579 by dropping the ts == 0 behavior from traceGoSysExit. Effectively restored, but at the one call site that needs it.

This CL fixes a bug present before CL 12579, namely that if the time stamp was from before the trace started, the event would be dropped, causing the same kind of failure. This CL handles that case by choosing a later time stamp. This could only happen if the system call exit happened with tracing enabled and then tracing was disabled and reenabled before execute was finally called. We saw this happen just once in thousands of runs under stress.

Patch-set: 1