Update patch set 1
Patch Set 1:

With a fair amount of awful code inserted into the runtime at the right places, we were able to reliably reproduce the stack barrier panic, and we confirmed that this change does fix it.

Rather than pollute the runtime to try to coerce it into failing in a test, the change to goexit's signature makes sure no one ever calls it accidentally again.

Patch-set: 1