[dev.ssa] cmd/compile: very first cut at adding FP support
Added F32 and F64 load, store, and addition.

I'm fishing for any comments on this early work to see if
the general direction is correct before I pile on a lot more
changes.  I did a fair amount of monkey-see-monkey-do.  It
passes one trivial test where I checked the assembly
language output and result of running the program, and it
passes all.bash.

Change-Id: Ifa60dbad948a40011b478d9605862c4b0cc9134c