cmd/compile: remove stale register use array
The reg[] array in .../gc is where truth lies.  The copy in .../amd64
is incorrect as it is mostly not updated to reflect regalloc decisions.

This bug probably also affects other archs.  I will add those after
a round of review.

This bug was introduced in the rewrite
https://go-review.googlesource.com/#/c/7853/.  The new reg[] array was
introduced in .../gc but not all of the uses were removed in the
.../ARCH directories.

Fixes #12133

Change-Id: I6364fc403cdab92d802d17f2913ba1607734037c