cmd/compile/internal/gc: compact binary export format
For #6110.
For #1909.

Status:
- compiles std/lib: (export GO_GCFLAGS=-newexport; make.bash)
- matching go/types importer implemented
- first round of improving data density
- 3 failures in all.bash

Missing:
- escape information
- inlined functions
- cleanups

Some initial export data sizes for top-level packages:

package         new     old	new/old

bufio.a		1641	7657	21.43%
bytes.a		2375	10346	22.96%
crypto.a	610	1840	33.15%
encoding.a	213	387	55.04%
errors.a	37	278	13.31%
expvar.a	855	3159	27.07%
flag.a		2254	19864	11.35%
fmt.a		959	3141	30.53%
hash.a		258	595	43.36%
html.a		71	221	32.13%
image.a		4383	27860	15.73%
io.a		2161	6769	31.92%
log.a		2294	11624	19.74%
math.a		2243	7515	29.85%
mime.a		1110	5601	19.82%
net.a		10283	43922	23.41%
os.a		5109	24879	20.54%
path.a		199	782	25.45%
reflect.a	3954	15473	25.55%
regexp.a	3651	13631	26.78%
runtime.a	1727	13616	12.68%
sort.a		696	2852	24.40%
strconv.a	832	2594	32.07%
strings.a	1727	7987	21.62%
sync.a		767	2572	29.82%
syscall.a	36536	81256	44.96%
testing.a	3770	16506	22.84%
time.a		2788	12107	23.03%
unicode.a	3202	9710	32.98%

total		96705	354744	27.26%
total -syscall	60169	273488	22.00%

Change-Id: Id657334f276383ff1e6fa91472d3d1db5a03349c