Update patch set 4
Patch Set 4:

This is the second cycle that this patch won't go in, because Marko got stalled on reviews. 

Last time I asked if anything's needed to make sure this patch lands, I received a rather glib response along the lines of "That would go into Go 1.5 anyway, not Go 1.4, so it can't really be blocking you more than awaiting Go 1.5 itself?"

It is as it is, but for what it's worth, this patch is *still* stalled on reviews.

Patch-set: 4
Reviewer: Gerrit User 6171 <6171@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0