Update patch set 1
Patch Set 1: Code-Review+1

Prior to this change 
 
    go test -v -run=TestCgoCallbackGC 

Would not pass on a Raspberry Pi 2, even with no other activity on the machine. With this change the test passes reliably, although I am still stress testing this change.

The values of 100 concurrent threads could be further reduced to improve this test on linux/arm if required.

Patch-set: 1