runtime: (SUPER HACKY) refactor runtime with the sockdrawer
DO NOT SUBMIT/DO NOT REVIEW

This is an attempted refactoring of the runtime package using
adonovan's sockdrawer tool.  I picked the runtime.clusters file in the
cl produced the runtime package and made changes to try to get the
refactored runtime package to compile.  For simplicity the refactored
runtime package only contains code for darwin/amd64. This should be
easy to fix the refactoring is automated.

The code in this CL is not very pretty.  I did the bare minimum to
satisfy compiler errors.

Current status (update as of 20 Aug 2015) I've been working on getting
the refactoring up-to-date with changes made to the runtime since
February. It's a little behind though -- the current revision can
build the bootstrap cmd/go, but it immediately crashes. Like the
previous versions, almost all the changes are automated on top of a clean
runtime.

Previous status (update as of 9 Feb 2015) cmd/go compiles and runs
but some tests fail.  Almost all changes needed to get to a compile
on darwin/amd64 are automated based on the config.  I'm currently
testing out the automation on a newer commit version, and fixing the
issues introduced by the differences.

See golang.org/cl/3269 for the tools used to produce the refactoring
and https://codereview.appspot.com/186270043/ for adonovan's original
sockdrawer CL.

Change-Id: Icdbe3fa54ce16c78acb6d84490463cc203a76b34