crypto/tls: fixed bug in tls.Listen where it was required to have a certificate
In go 1.5 it is allowed to have Config.Certificates nil as long as Config.GetCertificate 
is not also nil. Both cannot be nil. Just made sure the if statement in tls.Listen only returned a error
if config is nil OR (length of Certificates is 0 AND GetCertificate is nil)

I also submitted another change regarding the comment for this function
stating that atleast one certificate is required.

Change-Id: I5f67f323f63c988ff79642f3daf8a6b2a153e6b2