encoding/json: use bufio.Writer
Use a bufio.Writer whenever possible instead of a bytes.Buffer.

bytes.Buffers are still required in a few places, so use a pool.

The Encoder flushes after each Encode(). It may be useful to instead
provide a Flush() method which would proxy to the underlying
encodeState.

A Close() method on Encoder would allow for recycling of the
encodeState.

Both of those option though increase the size of Encoder's API,
so I didn't implement them.

For initial review.

Fixes #7872

Change-Id: I0088c09c6d96bfc01b8ab9ac0c194293b05e8dc4