cmd/asm: fix potential infinite loop in parser
For ARM machines, the assembler supports list of registers
operands such as [R1,R2].

An incomplete list (with a missing ]) results in the
parser issuing many errors and consuming all the tokens.
At EOF, it still loops.

Normally, a counter is maintained to make sure the parser
stops after 10 errors. However, multiple errors occuring on the
same line are simply ignored. Only the first one is reported.
At most one error per line is accounted.

An incomplete list of registers at EOF therefore results in an
infinite loop.

Fixed the parser by explicitly checking incomplete lists of registers.

In the operand tests, also fixed a wrong entry which I think was
not set on purpose (but still led to a successful result).

Fixes #11764

Change-Id: Ie87773388ee0d21b3a2a4cb941d4d911d0230ba4