Update patch set 4
Patch Set 4:

(5 comments)

Using relro only works if the linker is invoked with -z relro.  That's the default for Ubuntu GCC, and probably Debian too, so you may not have noticed.  But I think we ought to explicitly pass -Wl,-z,relro to the host linker.  And of course this only makes sense on systems that support that.  We wouldn't want to be doing this any of this folderol on Windows or Darwin, so I think the code needs some more conditionals.

Patch-set: 4
Label: Code-Review=0