Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Thanks for this.
 > 
 > I guess this could be tested in shared_test.go, not sure if it's
 > worth it.

Yes, I thought about these checks, and it is not difficult to implement them. However, what to do if we building a library like this:
go install -buildmode=shared -linkshared ./ ../splay/
it will generate 
'lib.-,..-splay-.so' which does not look good as well, despite the fact that it works.
I would like to implement the best variant in this case, but I am not sure which exactly to choose.
Probably, it is better to:
1) use explicit arguments without '.' and '/' if they are given
2) import path if arguments after filtering become empty ?
Thanks.

Patch-set: 1
Label: Code-Review=0