cmd/go: fixes Go buildid reading on Solaris
TestNoteReading fails on Solaris with linkmode=external due to some
assumptions made about how ELF .note sections are written by some linkers.

On current versions of Solaris and older derivatives, SHF_ALLOC is
intentionally ignored for .note sections unless the .note section is
assigned to the text segment via a mapfile.  Also, if .note sections are
assigned to the text segment, no PT_NOTE program header will be created
thwarting Go's attempts at attempting to quickly find the .note.

Furthermore, Go assumes that the relevant note segment will be placed early
in the file while the Solaris linker currently places the note segment last
in the file, additionally thwarting Go's optimisation attempts that read
only the first 16KB of the file to find the buildid.

The fix is to fallback to reading the entire file on Solaris if the buildid
note can't be found and fail only then if it still can't be found.  This
way, in future versions of Solaris when this linking behaviour is changed,
the fast path will always succeed and we'll only be slower if it fails.

Fixes #12178

Change-Id: I61c1dc3f744ae3ad63938386d2ace8a432c0efe1