Update patch set 1
Patch Set 1:

> I'd much rather see this check for an IsZero method and delegate to
 > that, which works for time.Time but also permits any other type to
 > work with this, not just structs.

That's the second part of the proposal in #11939, but I haven't submitted a CL for that yet.  I'll do that tomorrow.  There is some backward compatibility considerations with that, but I'm not sure if it's a problem.  I was hoping for some opinions in the issue from the core committers about that.

Worst case we could do that only in the case of structs, which up to this point have not supported omitempty at all, and don't have backward compatibility concerns.

Patch-set: 1