Update patch set 1
Patch Set 1:

I was going under the assumption that tests get checked in with the change. However, the following release schedule will avoid breakages at certain intermediate points:

Split this CL in two: 1) the changes to reflect minus the tests and 2) the gc change and tests.

Then:
1. Check in the changes to reflect. This will complete the changes for gccgo.
2. Check in fixes for package xml and json. This is the point also where fixes for other packages, with proper tests, can be committed.
3. Check in the changes to gc and the tests of this CL. At this point gccgo and gc are compatible.

Patch-set: 1
Label: Code-Review=0