Update patch set 1
Patch Set 1:

(2 comments)

New patch coming.  Discovered, after addressing Todd's remarks, that complex division is done in the runtime anyway, so the code I wrote is currently not tested.

I hope like heck that nobody is depending on catastrophic cancellation errors in complex arithmetic, and I think we should work as fast as possible to get any glitches like that out of the implementation ASAP so that people won't start to depend on them.

Patch-set: 1
Label: Code-Review=0