Update patch set 2
Patch Set 2: -Code-Review

Dammit, I did not mean to hit that button.

Josh, I don't think we really have much choice.  Bad cancellation in complex arithmetic is a known-stupid, and if someone is so gung-ho for performance that they'll live with it, they can pretty easily implement that version themselves.  The default behavior ought to be sane, and if the current behavior is not sane, the longer we wait, the deeper the hole we'll have dug.  I'll check the old code tomorrow.

Patch-set: 2