Update patch set 2
Patch Set 2:

So, the algorithm for import path's is:
Use arguments for special packages:
	std --> libstd.so
	std cmd --> libstd,cmd.so
	
Use import path for other cases:
	somelib --> libsubdir-somelib.so - exact package's import path
	somelib somelib2 --> libsubdir-somelib,subdir2-somelib-2.so - exact package's import path
	./ or ../ --> libsubdir-somelib.so - exact package's import path
	./... ---> libpkg1,pkg2.so - subset of all the import path's
If there are too many packages to build, it is better to limit the len of the
name of the library. So, there is no need determine a package(s) for every argument.
What do you think?
Thanks.

Patch-set: 2
Label: Code-Review=0