Update patch set 1
Patch Set 1:

> Sorry, this failed on my arm64 system, http://paste.ubuntu.com/12235595/

I'm pretty sure this is the result of a binutils bug (https://sourceware.org/ml/binutils/2015-01/msg00144.html). I'll try to get the fix backported to trusty, but I'm not sure what to do in the meantime to let us test this on the builders.

Patch-set: 1
Label: Code-Review=0