Update patch set 2
Patch Set 2:

Split the CL in this one and https://go-review.googlesource.com/14085.

After this change is merged, gccgo will be fully confirm the new semantics and can act as a reference implementation.

By removing the tests and the compiler change, this CL can be checked in without breaking anything. This at least solves the immediate issue that xml and json don't work well with gccgo (after checking in the respective CLs) and allows other package maintainers with similar issues to get their packages working for both compilers now.

After a grace period, we can make gc adopt the same semantics as gccgo.

Patch-set: 2