Update patch set 2
Patch Set 2:

One way to make it "test-only" would be to enable it via a compiler flag, and then arrange for 'go test' to pass that flag. That'd probably also require teaching test/run.go and a couple of other places about the flag. (All that said, I think it'd be better without being test-only.)

Patch-set: 2
Label: Code-Review=0