Update patch set 1
Patch Set 1:

(4 comments)

> (4 comments)
 > 
 > This is not the patch you showed me earlier, the one with the
 > linker mapfile, please upload that one instead. Thanks.

Unfortunately, the earlier version I showed you is not a viable solution.  It triggers a bug in all known versions of the Solaris linker that causes an invalid program header to be generated with some executables resulting in immediate segfault during linker setup at execution.

As a result, I have exhausted all possible options for controlling the layout of the elf file with the external linker on Solaris, the only remaining option is to either fallback to reading the entire ELF file, or to detect when the discovered note segment is outside of the first 16KB and read only that section via modification of readELFGoBuildID.

Pleas advise.

Patch-set: 1