Update patch set 1
Patch Set 1:

An alternative to consider: Use AuxInt to store float values instead of Aux. Then CSE will get identity right, as will any other passes. This would also reduce allocations. It's not obvious to me whether that would be better or worse than this fix. Thoughts? Could also be a todo.

As for this fix, please check whether AuxInt is nonzero and fail if so, to avoid clobbering it in (theoretical future) ops with both an auxint and a float aux. (Comments in context are impossible on a phone. Sorry and grrr.)

Patch-set: 1