Update patch set 1
Patch Set 1:

I was planing to do it differently. I was going to add new Go test that builds Go executable (using both internal and external linker). And check that it has appropriate symbol in its symbols table and it is in the right section and its address is aligned. Just for windows. We, probably, already have most of this in debug/pe.

Patch-set: 1