Update patch set 1
Patch Set 1:

The change itself looks good to me. I thought the eligibility criterion for a port on a release branch was the fact the problem is blocking for some users. I doubt users will be blocked by this bug (they just have to fix their assembler code to avoid it).

Patch-set: 1