Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > > Unfortunately, the earlier version I showed you is not a viable
 > > solution.  It triggers a bug in all known versions of the Solaris
 > > linker that causes an invalid program header to be generated with
 > > some executables resulting in immediate segfault during linker
 > setup
 > > at execution.
 > 
 > Could you provide more details about the bug? What executables are
 > problematic? Why? Can we work around it by generating different
 > executables?
 > 
 > What about using a different linker map, can we avoid the bug that
 > way?

Not that I'm aware of, to fit Go's constraints of placing the note segment within the first 16KB, this was the only option discovered after almost two weeks of different experiments.

 > What about fixing the bug in Solaris?

The crash is a bug, but it affects all known versions of Solaris, including older OpenSolaris-based derivatives.

Regardless, even if the segfault bug was fixed, I've been informed that the linker ordering in the mapefile I had that I was relying on was also considered a bug so should not be used.

 > > As a result, I have exhausted all possible options for
 > controlling
 > > the layout of the elf file with the external linker on Solaris,
 > the
 > > only remaining option is to either fallback to reading the entire
 > > ELF file, or to detect when the discovered note segment is
 > outside
 > > of the first 16KB and read only that section via modification of
 > > readELFGoBuildID.
 > 
 > Yes, let's only read what is needed.
 > 
 > I am very sad that Solaris has to take this performance hit
 > compared
 > to other systems. You said we can't put the build id in the text

At some future date, the Solaris linker and hopefully other older, OpenSolaris-based derivatives will be changed to more closely match the GNU linker behaviour for output placement within the file.

What the Solaris linker does today (excluding the crash with the mapfile) is expected and acceptable according to the ABI.  What Go does today with note sections relies on undocumented implementation-specific behaviour.  (The assumption that notes are in the first 16KB.)

 > segment, because on Solaris the text segment is not the first one
 > in the binary. Well, what is the first thing in the binary, can't
 > we put it there?

The first sections are those generated by the Solaris linker itself; program data and text comes after that and notes are last.

I'm not aware of any reliable way of placing anything in those sections since they're generated by the linker.

The Solaris linker currently offers no viable way to control output order of sections written to the ELF file.

Patch-set: 1